refactor: move user fetching logic to RootLayout for better routing
- Moved user fetching and routing logic from AuthRoutesLayout to RootLayout.
- Ensures users are redirected to the correct page (home or onboarding) upon app entry.
